addressed jeffvance's review comments
